pos-edi/infra/airflow/requirements.txt
mysql-connector-python
psycopg2-binary
pandas
